moved GetConfigOptions API view to Poem/api/internal_views/app.py file
