+++ poem/Poem/api/internal_views/app.py
@@ -1,4 +1,7 @@
+from django.conf import settings
+
 from Poem.api.internal_views.users import get_all_groups, get_groups_for_user
+from Poem.poem.saml2.config import tenant_from_request, saml_login_string
 
 from rest_framework.authentication import SessionAuthentication
 from rest_framework.response import Response
@@ -21,3 +24,20 @@ class ListGroupsForUser(APIView):
             return Response(results[group.lower()])
         else:
             return Response({'result': results})
+
+
+class GetConfigOptions(APIView):
+    authentication_classes = ()
+    permission_classes = ()
+
+    def get(self, request):
+        options = dict()
+
+        tenant = tenant_from_request(request)
+        options.update(saml_login_string=saml_login_string(tenant))
+
+        options.update(webapimetric=settings.WEBAPI_METRIC)
+        options.update(webapiaggregation=settings.WEBAPI_AGGREGATION)
+        options.update(tenant_name=tenant)
+
+        return Response({'result': options})
